+++ bitcampfire/src/main/java/com/modak/user/controller/UserSignupController.java
@@ -96,11 +96,10 @@ public class UserSignupController {
 	        String toMail = user_email;
 	        String title = "회원가입 인증 이메일 입니다.";
 	        String content = 
-	                "홈페이지를 방문해주셔서 감사합니다." +
-	                "<br><br>" + 
-	                "인증 번호는 " + checkNum + "입니다." + 
-	                "<br>" + 
-	                "해당 인증번호를 인증번호 확인란에 기입하여 주세요.";
+	        "<div style='width:1000px; height: 100px; background:#286090;' align='center'> <h1 style='color:#fff; font-size: 60px;'>BITFIRE</h1></div>"
+			+ "<div><h2 style='margin-top:10px; font-size: 24px;'>홈페이지를 방문해주셔서 감사합니다.<br><br>아래의 인증번호를 인증번호 확인란에 기입하여 주세요.</h2>"
+			+ "<p style='font-size:18px;'>인증 번호 : <span style='padding: 10px; background: #d2e9fc;'>" + checkNum + "</span> </p>"
+			+ "<div>";
 	        
 	        try {
 	            
@@ -140,48 +139,7 @@ public class UserSignupController {
 		
 		
 		
-		//이메일 계정찾기 전송
 		
-		@GetMapping("pwdFindmailCheck")
-		@ResponseBody
-		public String pwdFindmailCheck(String user_email) throws Exception{
-			logger.info("이메일 인증 요청이 들어옴"+user_email);
-	        logger.info("인증번호 : " + user_email);
-			//return  mailService.joinEmail(user_email);
-	        
-	        /* 인증번호(난수) 생성 */
-	        Random random = new Random();
-	        int checkNum = random.nextInt(888888) + 111111;
-	        logger.info("인증번호 " + checkNum);
-	        
-	        /* 이메일 보내기 */
-	        String setFrom = "yujin980810@gmail.com";
-	        String toMail = user_email;
-	        String title = "회원가입 인증 이메일 입니다.";
-	        String content = 
-	                "홈페이지를 방문해주셔서 감사합니다." +
-	                "<br><br>" + 
-	                "인증 번호는 " + checkNum + "입니다." + 
-	                "<br>" + 
-	                "해당 인증번호를 인증번호 확인란에 기입하여 주세요.";
-	        
-	        try {
-	            
-	            MimeMessage message = mailSender.createMimeMessage();
-	            MimeMessageHelper helper = new MimeMessageHelper(message, true, "utf-8");
-	            helper.setFrom(setFrom);
-	            helper.setTo(toMail);
-	            helper.setSubject(title);
-	            helper.setText(content,true);
-	            mailSender.send(message);
-	            
-	        }catch(Exception e) {
-	            e.printStackTrace();
-	        }
-	        String num = Integer.toString(checkNum);
-	        
-	        return num;
-		}
