+++ bitcampfire/src/main/resources/spring/db.properties
@@ -1,4 +1,4 @@
 jdbc.driver=oracle.jdbc.driver.OracleDriver
 jdbc.url=jdbc:oracle:thin:@localhost:1521:xe
-jdbc.username=java
-jdbc.password=bit
+jdbc.username=c##java
+jdbc.password=oracle
